+++ modules/main/src/main/java/com/nisovin/shopkeepers/shopobjects/block/base/BaseBlockShopListener.java
@@ -1,6 +1,9 @@
 package com.nisovin.shopkeepers.shopobjects.block.base;
 
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
+import java.util.stream.Stream;
 
 import org.bukkit.Bukkit;
 import org.bukkit.World;
@@ -21,7 +24,6 @@ import org.bukkit.event.entity.EntityExplodeEvent;
 import org.bukkit.event.player.PlayerInteractEvent;
 import org.bukkit.inventory.EquipmentSlot;
 import org.checkerframework.checker.nullness.qual.NonNull;
-import org.checkerframework.checker.nullness.qual.Nullable;
 
 import com.nisovin.shopkeepers.SKShopkeepersPlugin;
 import com.nisovin.shopkeepers.api.internal.util.Unsafe;
@@ -30,6 +32,7 @@ import com.nisovin.shopkeepers.config.Settings;
 import com.nisovin.shopkeepers.shopkeeper.AbstractShopkeeper;
 import com.nisovin.shopkeepers.shopkeeper.registry.SKShopkeeperRegistry;
 import com.nisovin.shopkeepers.util.bukkit.BlockFaceUtils;
+import com.nisovin.shopkeepers.util.bukkit.BlockLocation;
 import com.nisovin.shopkeepers.util.bukkit.EventUtils;
 import com.nisovin.shopkeepers.util.bukkit.MutableBlockLocation;
 import com.nisovin.shopkeepers.util.bukkit.TextUtils;
@@ -39,17 +42,37 @@ import com.nisovin.shopkeepers.util.logging.Log;
 
 class BaseBlockShopListener implements Listener {
 
-	// Local copy as array (enables a very high-performance iteration):
+	// Local copy as array: Enables high-performance iteration.
 	// This includes all directions that physic updates can propagate from, regardless of whether
 	// the block shops can be attached in that direction.
 	private static final @NonNull BlockFace[] BLOCK_SIDES = BlockFaceUtils.getBlockSides()
 			.toArray(new @NonNull BlockFace[0]);
 
+	// The block faces that physic updates can propagate from and potentially affect nearby blocks.
+	// Spigot changed the behavior of the physics event in MC 1.13 to reduce the number of event
+	// calls (https://hub.spigotmc.org/jira/browse/SPIGOT-4256). We therefore need to check the
+	// neighboring blocks ourselves.
+	private static final @NonNull BlockFace[] PHYSICS_BLOCK_FACES
+			= Stream.concat(Stream.of(BlockFace.SELF), BlockFaceUtils.getBlockSides().stream())
+					.toArray(length -> new @NonNull BlockFace[length]);
+
+	private static final MutableBlockLocation SHARED_BLOCK_LOCATION = new MutableBlockLocation();
+
 	private final SKShopkeepersPlugin plugin;
 	private final BaseBlockShops baseBlockShops;
 	private final SKShopkeeperRegistry shopkeeperRegistry;
 
-	private final MutableBlockLocation cancelNextBlockPhysics = new MutableBlockLocation();
+	// We cancel all physics events around block shops in order to e.g. prevent sign shops from
+	// breaking.
+	// For performance optimization purposes, instead of checking during each physics event whether
+	// the event shall be cancelled for the block or one of its potentially affected neighbors
+	// (7 block checks), we pre-calculate in this map the block locations for which physics events
+	// shall be cancelled and then only do a single block lookup during the event.
+	// Value: The number of "tickets" asking for block physics to be cancelled at the specific block
+	// location. E.g. the number of block shopkeepers, but tickets can also be added for other
+	// purposes. When this number reaches zero, the map entry is removed and block physics are no
+	// longer cancelled for the location.
+	private final Map<BlockLocation, Integer> cancelledBlockPhysics = new HashMap<>();
 
 	BaseBlockShopListener(SKShopkeepersPlugin plugin, BaseBlockShops blockShops) {
 		this.plugin = plugin;
@@ -70,13 +93,60 @@ class BaseBlockShopListener implements Listener {
 		HandlerList.unregisterAll(this);
 	}
 
-	// Null to clear
-	void cancelNextBlockPhysics(@Nullable Block block) {
-		if (block == null) {
-			cancelNextBlockPhysics.setWorldName(null);
-		} else {
-			cancelNextBlockPhysics.set(block);
+	void addBlockPhysicsCancellation(Block block) {
+		var blockLocation = SHARED_BLOCK_LOCATION;
+		blockLocation.set(block);
+		this.addBlockPhysicsCancellation(blockLocation);
+	}
+
+	void removeBlockPhysicsCancellation(Block block) {
+		var blockLocation = SHARED_BLOCK_LOCATION;
+		blockLocation.set(block);
+		this.removeBlockPhysicsCancellation(blockLocation);
+	}
+
+	void addBlockPhysicsCancellation(BlockLocation blockLocation) {
+		for (BlockFace blockFace : PHYSICS_BLOCK_FACES) {
+			int adjacentX = blockLocation.getX() + blockFace.getModX();
+			int adjacentY = blockLocation.getY() + blockFace.getModY();
+			int adjacentZ = blockLocation.getZ() + blockFace.getModZ();
+			var adjacentBlockLocation = new BlockLocation(
+					Unsafe.assertNonNull(blockLocation.getWorldName()),
+					adjacentX,
+					adjacentY,
+					adjacentZ
+			);
+			this.addSpecificBlockPhysicsCancellation(adjacentBlockLocation);
 		}
+	}
+
+	void removeBlockPhysicsCancellation(BlockLocation blockLocation) {
+		for (BlockFace blockFace : PHYSICS_BLOCK_FACES) {
+			int adjacentX = blockLocation.getX() + blockFace.getModX();
+			int adjacentY = blockLocation.getY() + blockFace.getModY();
+			int adjacentZ = blockLocation.getZ() + blockFace.getModZ();
+			var adjacentBlockLocation = new BlockLocation(
+					Unsafe.assertNonNull(blockLocation.getWorldName()),
+					adjacentX,
+					adjacentY,
+					adjacentZ
+			);
+			this.removeSpecificBlockPhysicsCancellation(adjacentBlockLocation);
+		}
+	}
+
+	private void addSpecificBlockPhysicsCancellation(BlockLocation blockLocation) {
+		cancelledBlockPhysics.compute(
+				blockLocation.immutable(),
+				(k, v) -> v == null ? 1 : (v + 1)
+		);
+	}
+
+	private void removeSpecificBlockPhysicsCancellation(BlockLocation blockLocation) {
+		cancelledBlockPhysics.compute(
+				blockLocation.immutable(),
+				(k, v) -> (v == null || v <= 1) ? null : (v - 1)
+		);
 	}
 
 	// See LivingEntityShopListener for the reasoning behind using event priority LOWEST and
@@ -143,6 +213,9 @@ class BaseBlockShopListener implements Listener {
 
 	// Protect shop blocks:
 
+	// TODO Pre-calculate the set of protected blocks similar to how we pre-calculate the set of
+	// block locations affected by block physics? We would need to also take the attached block face
+	// into account.
 	private boolean isProtectedBlock(Block block) {
 		// Check if the block itself is a base block shop:
 		if (baseBlockShops.isBaseBlockShop(block)) {
@@ -203,32 +276,13 @@ class BaseBlockShopListener implements Listener {
 		int blockX = block.getX();
 		int blockY = block.getY();
 		int blockZ = block.getZ();
-		if (this.checkCancelPhysics(worldName, blockX, blockY, blockZ)) {
-			event.setCancelled(true);
-			return;
-		}
-		// Spigot changed the behavior of this event in MC 1.13 to reduce the number of event calls:
-		// Related: https://hub.spigotmc.org/jira/browse/SPIGOT-4256
-		for (BlockFace blockFace : BLOCK_SIDES) {
-			// Note: Avoiding getting the adjacent block slightly improves performance of handling
-			// this event.
-			int adjacentX = blockX + blockFace.getModX();
-			int adjacentY = blockY + blockFace.getModY();
-			int adjacentZ = blockZ + blockFace.getModZ();
-			if (this.checkCancelPhysics(worldName, adjacentX, adjacentY, adjacentZ)) {
-				event.setCancelled(true);
-				return;
-			}
-		}
-	}
 
-	private boolean checkCancelPhysics(String worldName, int blockX, int blockY, int blockZ) {
-		if (cancelNextBlockPhysics.matches(worldName, blockX, blockY, blockZ)) {
-			return true;
-		} else if (baseBlockShops.isBaseBlockShop(worldName, blockX, blockY, blockZ)) {
-			return true;
+		var blockLocation = SHARED_BLOCK_LOCATION;
+		SHARED_BLOCK_LOCATION.set(worldName, blockX, blockY, blockZ);
+
+		if (cancelledBlockPhysics.containsKey(blockLocation)) {
+			event.setCancelled(true);
 		}
-		return false;
 	}
 
 	@EventHandler(priority = EventPriority.LOW, ignoreCancelled = true)